Versioned Storages (#1805) - Resolve conflicts: Binary file form
