+++ client/src/models/basic/Remote.js
@@ -68,6 +68,10 @@ class Remote {
     return this._response;
   }
 
+  @observable responseStatus = undefined;
+  @observable responseStatusText = undefined;
+  @observable responseError = false;
+
   _loadRequired = !this.constructor.auto;
 
   async _fetchIfNeeded () {
@@ -105,6 +109,12 @@ class Remote {
           }
           fetchOptions.headers = headers;
           const response = await fetch(`${prefix}${this.url}`, fetchOptions);
+          this.responseError = !response.ok;
+          this.responseStatus = response.status;
+          this.responseStatusText = response.statusText;
+          if (!response.ok) {
+            // throw new Error(response.statusText || `HTTP Error ${response.status}`);
+          }
           const data = this.constructor.isJson ? (await response.json()) : (await response.blob());
           this.update(data);
         } catch (e) {
